Refactoring and continuing to work on 'suitable for {diets}' addition to recipe cards
