+++ src/App.css
@@ -823,6 +823,10 @@ h4 {
   font-weight: 500;
 }
 
+.diet {
+  text-align: left;
+}
+
 .additionalIngredients__list p,
 .usedIngredients__container p {
   margin: 0;
